refactor: rename portfolio domain directory to layout
- Rename src/domains/portfolio to src/domains/layout for better semantic clarity
- Update all import paths to reference new layout directory structure
- Maintain existing functionality and component organization
