+++ src/domains/layout/components/TechnicalImplementationSection/TechnicalImplementationSection.tsx
