+++ src/domains/layout/utils/skillExtraction.ts
