+++ src/scripts/analyzeSkillExtraction.ts
@@ -3,7 +3,7 @@
  */
 
 import { recommendations } from '../domains/recommendations/data/recommendations';
-import { analyzeSkillExtractionOpportunities, extractSkillsFromContent } from '../domains/portfolio/utils/skillExtraction';
+import { analyzeSkillExtractionOpportunities, extractSkillsFromContent } from '../domains/layout/utils/skillExtraction';
 
 function runAnalysis() {
   console.log('🔍 Analyzing skill extraction opportunities...\n');
